Move more logic to parser and transformer
Previously the parser and transformer simply handled the type
and constraint information as text, leaving all logic to the
script calling them.

With this change, the parser breaks up the fixed length fields
information in lengths and constraints, and the transformer takes care
to map these to the expected output rules.

Unfortunately much of the type information is free text, so many hard
coded rules remain in the script. (Even some of the parsed constraints
are more or less hard coded, but at least they are handled in the
transformer.)

A few rules were restructured and reworded (e.g. using "natural"
instead of "non_zero_digit") and one was fixed; fee_limit was
incorrectly set to four digits, while actually being a variable length
field.