Add parser, transformer and script to test them
The parser is simply a root pointer into the messages_parser_rules.

The transformer handles the generic atoms generated by both
message_parser_rules and field_parser_rules, but also a few more complex
rules matching on `subtree` to beat tree into shape.

The script simply runs the parser and transformer on ARGF (STDIN or
named file) and dumps the result as JSON. Note that the complete input
is read and then parsed, outputting all messages in a (potentially
large) array. It might be desirable to read line by line and output a
JSON stream instead - that might also allow for better error handling on
a line by line basis.