Add option to read and parse input line by line
Passing "-1" as the first argument will now read one input line at a
time, parsing it and outputting the *unwrapped* message (the parser
produces an array of messages) as a JSON object. With several lines, the
complete output will not be valid JSON, but tools like `jq` can accept a
series of objects as input.