Include support for error detection fields
Error detection fields are labeled but has no delimiter. When enabled,
they appear last in a message, before the terminating carriage return.
Both a sequence number and a checksum must be present, except for a
couple of messages where only the checksum is allowed.

The position and nature of the error detection fields requires capturing
them to a new "group", to enable merging. This leaves a total of four
groups where only one must be present in all messages. To make the merge
transformation simpler, each message now produces all groups, even if
some are empty.