Accept messages with unknown message identifiers
The specification says that messages with unknown "command identifier"
should be ignored and that a "command identifier" consists of "two ASCII
characters".

The only two characters that are explicitly disallowed are NULL
(disallowed everywhere in a message) and carriage return (allowed only
as message terminator).

Existing messages all have two digit identifiers, and there is plenty of
room for more. I consider it highly unlikely that punctuation or control
characters will be used as message identifiers, and chose to allow only
digits or letters for now.