Add WIP SIP2 message printing support
Add script to generate a FIELDS constant in the form of a hash linking
field symbol names to a field code (if any), a Dry::Type, and a lambda
to format the value.

Add Checkout class as a Dry::Struct with helpers to populate a fields
list, generate attributes and to print a Sip2 representation of the
values. This class can serve as a template for which a builder can
generate structs for all message types.

Add script to round-trip checkout messages for testing.

CAVEAT: Timezones are currently not round-tripped correctly. UTC/Z
works. Blanks for using local timezone works. But any other parsed
timezone might not roundtrip - an explicit local timezone will be
replaced with blank, a "military timezone" will be replaced with a three
or four letter timezone or with blanks if none is found (resulting in an
erroneous time).