Generate redo scripts to create message struct classes
The generated classes all depend on a common `Message::BaseMessage`
class that adds helper methods `fields` and `fields?` that generate
`attributes` from definitions in `Sip2::FIELDS`, and stores the
attributes added this way in a class variable. The class variable is
used in another helper that maps each added field to it's writer.

Currently only top level attributes work, nested structs are not yet
implemented.

In the generated classes all fields are required, so the manually
crafted `Checkout` message can not yet be replaced with the generated
`Message::Checkout` in the round-trip.

A message.do script is provided that will generate all messages defined
in doc/sip2_messages.json.