Create message parser rules dynamically
Previously we had message codes, names and symbols in a constant, and
information on fields encoded in the composed message rules - originally
created through code generation. By moving the field information into
the constant we can dynamically create the rules from this list.

If we also reuse the list for dynamic generation of the message structs
(currently code generated from JSON) we could have a "single
source of truth" for message formats.

This commit includes the build scripts and marks the messages list as
generated for now. But as it contains *all* data from the original JSON
file, in a next step we should be able to use the new constant as the
primary source.