Handle all optional fields in sample input
With this change, the round-trip script works for all messages in the
(not-included) sample file. Field order is not preserved so a standard
diff shows many differences, but running `sip2-to-json` on both the
input and the output of the round-trip and then diffing with `jd`
("jd-diff-patch") produces no semantic difference.

NOTE: Several messages remain where we haven't yet checked which fields
are optional and which are required.