Handle repeatable top level fields
"Print line" and "Screen message" can be repeated (and not necissarily
in direct order) resulting in messages of multiple lines. When merging
the list of messages, these fields must not overwrite earlier instances,
but together form a list of values.

TODO: The solution is *okay* but not great. However, the items subfields
use *a different* solution for repeated fields, that is worse. Even
though there are differences we could probably align the solutions a bit
more.