Add option to pre-process messages before checksumming
The SIP2 standard uses vague language on the encoding of messages, and
is especially unclear when it comes to encoding and the checksumming
algorithm.

The current naive implementation uses the code points of the incoming
message as read, i.e. probably UTF-8. This however is not necessarily
what you want, depending on what you are integrating with.

When calling Sip2.encode, you now have an option to pass a
`checksum_encoder`, a callable (e.g. proc) that accepts a string and
returns a re-encoded string. The actual message will keep the original
encoding, but when calculating the checksum, the result of passing the
message through the checksum_encoder will be used.

`Message.encode` accepts and uses this new parameter, and `Message.to_s`
now simply calls `Message.encode` without the parameter set.