Remove _fixed suffix from ordered variant attributes
Some fields exists both as ordered fields and as delimited fields for
different messages. Previously this was handled through a "_fixed"
suffix on the field name, and this suffix was present on the attribute
in the Message, and in JSON output.

With this change, the suffix on the field names has been replaced by
"_ordered", to better represent the difference from the non-suffixed
version of the field.  More importantly a field definition might now
have a :name key, which if set the value will be used for the attribute
name, and in the JSON output. For the "_ordered" versions of fields,
:name is set to the field name without suffix, and the rules for the
"_ordered" versions of the fields have also been updated to use this
name.

(Of the two fields, fee_type_ordered and currency_type_ordered, only
fee_type_ordered currently exists in the fixtures and is properly tested.)