Add support for unrecognized fields
The Sip2 standard states that an implementation should ignore fields
that it does not recognize. To allow pass through of messages we instead
collect them into it's own field in the Hash/JSON representation, and
back again to their original form on encoding.

An option is added to Sip2::BaseMessage.encode that allows the
unrecognised fields to be stripped out before encoding, and this option
is also added to json-to-sip2.