Fix time zone bug and document time zone limitations
Previously time stamps were incorrectly encoded using time zone names
from the time zone object, which would normally be in ISO8601 format.
The Sip2 standard specifies that time data should be in the ancient
standard ANSI X3.43 which uses ANSI X3.51 time zones, for which there is
limited support in Ruby.

Now all time zones are stripped when encoding, except UTC which is
always encoded as "Z". A caveat for this is added to the README.